Fix Travis Deploy
This should make Travis only try to deploy a Gem once, and avoid the
race condition encountered during the v1.1.0 release

This is a bit of a hack. To make sure that only one build tries to
deploy, I am setting a special flag during the GHP build, and then only
trying to deploy if that flag is set. This will prevent deploy during
any other Travis build.