Fix Jekyll version for GitHub Pages jobs
Previously, the GitHub Pages builds were not using the precise version
as specified in the pages-gem. The pages gem currently depends on
Jekyll v3.9.0.

Collateral damage of this commit is the removal of the git and
branches keys in the YAML that I will add back in when I submit a
PR upstream.